pyproject.toml
[project]
authors = [{ name = "Konovalov Dmitriy", email = "dkonowalow@gmail.com" }]
dependencies = [
  "fastapi>=0.111.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.1",
  "psycopg>=3.1.19",
  "psycopg-binary>=3.1.19",
  "pydantic>=2.7.1",
  "uvicorn>=0.20.0",
  "dishka>=1.2.0",
  "bcrypt>=4.2.0",
  "pyjwt>=2.8.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.9",
]
description = "URL shortener"
name = "url_shortener"
requires-python = ">=3.12"
version = "0.1.0"


[project.optional-dependencies]
dev = [
  "ruff>=0.4.4",
  "mypy>=1.9.0",
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.6",
  "httpx>=0.27.0",
  "pre-commit>=3.7.1",
]

[tool.pytest.init_options]
filterwarnings = ["ignore::DeprecationWarning"]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
exclude = "src/app/repositories/migrations"

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src"]
include = ["src/**.py", "tests/**.py"]
exclude = ["migrations**"]


[tool.ruff.lint]
ignore = [
  "ARG",
  "ANN",
  "D",
  "EM101",
  "EM102",
  "PT001",
  "PT023",
  "SIM108",
  "SIM114",
  "TRY003",
  "PLW2901",
  "RET505",
  "PLR0913",
  "UP038",
  "TCH001",
  "SLF001",
  "COM812",
  "ISC001",
  "S101",
]
select = ["ALL"]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"
